+++ A3D.Api/Startup.cs
@@ -1,28 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using A3D.Library.Repositories.EntityFramework;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Controllers;
 using Microsoft.AspNetCore.Mvc.ViewComponents;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 
 using SimpleInjector;
 using SimpleInjector.Lifestyles;
 using SimpleInjector.Integration.AspNetCore.Mvc;
+using A3D.Library.Models.LookUp;
+using A3D.Library.Repositories.EntityFramework;
+using A3D.Library.Repositories.EntityFramework.LookUp;
+using A3D.Library.Repositories.Interfaces;
 using A3D.Library.Services.Interfaces;
 using A3D.Library.Services;
-using A3D.Library.Repositories.LookUp;
-using A3D.Library.Models.LookUp;
+using A3D.Library.Services.LookUp;
 
 namespace A3D.Api
 {
