+++ A3D.Library/Repositories/Interfaces/ILookUpRepository.cs
@@ -1,7 +1,7 @@
 ﻿using A3D.Library.Models;
 using System.Linq;
 
-namespace A3D.Library.Repositories.LookUp
+namespace A3D.Library.Repositories.Interfaces
 {
     public interface ILookUpRepository<TEntity> where TEntity : BaseLookUpModel
     {
