+++ A3D.Library/Services/Interfaces/ILookUpService.cs
@@ -5,6 +5,10 @@ namespace A3D.Library.Services.Interfaces
 {
     public interface ILookUpService
     {
+        /// <summary>
+        /// Returns a dictionary of all {lookup-name, list-of-lookup-values}
+        /// </summary>
+        /// <returns></returns>
         IDictionary<string, IEnumerable<BaseLookUpModel>> GetAll();
     }
 }
